Add some stack corruption detection to tests
The API is wrapped with some code that tries to detect if things are
going wrong:

* Copies function pointer and arguments into callee-saved registers;
* Creates backup of new callee-saved registers;
* Creates large buffer from base-stack-pointer.

It checks that the buffer is in tact, then checks the register backups.
This results in confirming that we're returning from the correct
function to the correct place on the stack.

Hopefully, combined with a large battery of test cases (coming soon),
this should end up providing a complete set of safe test cases.