Add (theoretical) thread-safety to x86_64 (#7)
The current implementation passes existing tests, and while I'm pretty
sure it'll be thread-safe, I don't yet know how to test this. As there
is room for doubt, I am not yet documenting this behaviour.

This was started as an effort to clean away the unnecessary activate and
deactivate utility functions (in favour of calling _green_current
directly). Sort of got out of hand; I wasn't expecting to implement this
quite so early.

Since there are no tests for this, other architectures/platforms likely
won't mimic this capability right away.